Создал структуру папок компонентов
