+++ miniapp/src/components/CardComponents/CardList/CardList.module.scss
