+++ miniapp/src/components/SectionComponents/BlockLink/BlockLinkSection/BlockLinkSection.module.scss
