+++ miniapp/src/components/SectionComponents/ChooseSectionModal/ChooseSectionModal.module.scss
