+++ miniapp/src/components/SectionComponents/Divider/DividerSection/DividerSection.module.scss
