+++ miniapp/src/components/SectionComponents/Divider/DividerSection/DividerSection.tsx
