+++ miniapp/src/components/SectionComponents/Section/Section.module.scss
