+++ miniapp/src/components/SectionComponents/Text/TextSection/TextSection.module.scss
