+++ miniapp/src/pages/CardPage/CardPage.tsx
@@ -2,17 +2,17 @@ import React, {useEffect, useMemo} from 'react';
 import '@telegram-apps/telegram-ui/dist/styles.css';
 import '@mdxeditor/editor/style.css'
 import './mdxstyle.scss';
-import LinkModal from "../../components/LinkModal/LinkModal";
-import TextModal from "../../components/TextModal/TextModal";
-import VideoModal from "../../components/VideoModal/VideoModal";
-import ChooseSectionModal from "../../components/ChooseSectionModal/ChooseSectionModal";
-import CardPageBottomMenu from "../../components/CardPageBottomMenu/CardPageBottomMenu";
-import SectionList from "../../components/SectionList/SectionList";
+import LinkModal from "../../components/SectionComponents/BlockLink/LinkModal/LinkModal";
+import TextModal from "../../components/SectionComponents/Text/TextModal/TextModal";
+import VideoModal from "../../components/SectionComponents/Video/VideoModal/VideoModal";
+import ChooseSectionModal from "../../components/SectionComponents/ChooseSectionModal/ChooseSectionModal";
+import CardPageBottomMenu from "../../components/BottomMenu/CardPageBottomMenu/CardPageBottomMenu";
+import SectionList from "../../components/SectionComponents/SectionList/SectionList";
 import {useParams} from "react-router-dom";
 import {readOneBusinessCard} from "../../store/apiThunks/businessCardThunks";
 import {useAppDispatch, useAppSelector} from "../../hooks/hooks";
 import {getUserId} from "../../utils/getUserId";
-import CardPageBottomMenuForGuest from "../../components/CardPageBottomMenuForGuest/CardPageBottomMenuForGuest";
+import CardPageBottomMenuForGuest from "../../components/BottomMenu/CardPageBottomMenuForGuest/CardPageBottomMenuForGuest";
 import Loader from "../../components/Loader/Loader";
 import {closeAllModals} from "../../store/slices/modalsCardPageSlice";
 import {clearLinkSlice} from "../../store/slices/linkSlice";
@@ -20,8 +20,8 @@ import {clearTextSlice} from "../../store/slices/textSlice";
 import {clearVideoSlice} from "../../store/slices/videoSlice";
 import {selectSection} from "../../store/slices/myCardsSlice";
 import {CARD_MODE} from "../../constants/cardMode";
-import CardTitle from "../../components/CardTitle/CardTitle";
-import TitleModal from "../../components/TitleModal/TitleModal";
+import CardTitle from "../../components/SectionComponents/Title/CardTitle/CardTitle";
+import TitleModal from "../../components/SectionComponents/Title/TitleModal/TitleModal";
 
 const CardPage = () => {
     const dispatch = useAppDispatch();
