+++ miniapp/src/pages/CardsHistory/CardsHistory.tsx
@@ -1,12 +1,9 @@
 import React, {useEffect} from 'react';
 import '@telegram-apps/telegram-ui/dist/styles.css';
 import TopMenu from "../../components/TopMenu/TopMenu";
-import ModalAddCard from "../../components/ModalAddCard/ModalAddCard";
-import ModalDeleteCard from "../../components/ModalDeleteCard/ModalDeleteCard";
-import CardList from "../../components/CardList/CardList";
-import MyCardsBottomMenu from "../../components/MyCardsBottomMenu/MyCardsBottomMenu";
-import CardsHistoryList from '../../components/CardsHistory/CardsHistoryList';
-import CardsHistoryBottomMenu from '../../components/CardsHistoryBottomMenu/CardsHistoryBottomMenu';
+import ModalDeleteCard from "../../components/CardComponents/ModalDeleteCard/ModalDeleteCard";
+import CardsHistoryList from '../../components/CardsHistoryList/CardsHistoryList';
+import CardsHistoryBottomMenu from '../../components/BottomMenu/CardsHistoryBottomMenu/CardsHistoryBottomMenu';
 import {useAppDispatch, useAppSelector} from "../../hooks/hooks";
 import Loader from "../../components/Loader/Loader";
 import {getHistory} from "../../store/apiThunks/historyThunks";
