+++ miniapp/src/pages/MyCardsPage/MyCardsPage.tsx
@@ -1,11 +1,11 @@
 import React, {useEffect} from 'react';
 import '@telegram-apps/telegram-ui/dist/styles.css';
 import TopMenu from "../../components/TopMenu/TopMenu";
-import ModalAddCard from "../../components/ModalAddCard/ModalAddCard";
-import ModalDeleteCard from "../../components/ModalDeleteCard/ModalDeleteCard";
-import CardList from "../../components/CardList/CardList";
-import MyCardsBottomMenu from "../../components/MyCardsBottomMenu/MyCardsBottomMenu";
-import ModalSettingsCard from "../../components/ModalSettingsCard/ModalSettingsCard";
+import ModalAddCard from "../../components/CardComponents/ModalAddCard/ModalAddCard";
+import ModalDeleteCard from "../../components/CardComponents/ModalDeleteCard/ModalDeleteCard";
+import CardList from "../../components/CardComponents/CardList/CardList";
+import MyCardsBottomMenu from "../../components/BottomMenu/MyCardsBottomMenu/MyCardsBottomMenu";
+import ModalSettingsCard from "../../components/CardComponents/ModalSettingsCard/ModalSettingsCard";
 import {useAppDispatch, useAppSelector} from "../../hooks/hooks";
 import {readBusinessCards} from "../../store/apiThunks/businessCardThunks";
 import {useLocation} from "react-router-dom";
